fix(macos-ai-subscription-tracker): recover rejected overrides
